==> 03_interacao.qmd <==
---
title: "03: Adaptabilidade e estabilidade de genótipos experimentais de linhaça cultivados em ambientes contrastantes de Santa Catarina"
---

```{r global_options, include = FALSE}
knitr::opts_chunk$set(cache = FALSE,
                      comment = "##",
                      collapse = TRUE,
                      warning = FALSE,
                      message = FALSE)

```


# Pacotes

```{r warning=FALSE, message=FALSE}
library(rio)
library(tidyverse)
library(metan)
library(ggridges)
library(ggthemes)
library(patchwork)

df <- import_list("data/progresso.xlsx")
g2022 <- df$`2022` |> select(GEN, RGPLA) |> rename(A2022 = RGPLA)
g2023 <- df$`2023` |> select(GEN, RGPLA) |> rename(A2023 = RGPLA)
g2024 <- df$`2024` |> select(GEN, RGPLA) |> rename(A2024 = RGPLA)


dfge <- 
  reduce(list(g2024, g2023, g2022), left_join) |> 
  remove_rows_na()

df_ggplot <- 
  dfge |> 
  pivot_longer(-GEN, names_to = "AMB", values_to = "RGPLA")
```



# Interação genótipo ambiente

$$
y_{ij} = m + g_i + a_j + (ga)_{ij}
$$

$$
\hat{g_i} = \bar{g_i} - m  \\
\hat{a_j} = \bar{a_j} - m  
$$
$$
\hat{(ga)_{ij}} = y_{ij} - m - \hat{g_i} - \hat{a_j} \\
\hat{(ga)_{ij}} = y_{ij} - \bar{g_i} - \bar{a_j} + m
$$


```{r}
matga <- dfge |> column_to_rownames("GEN") |> as.matrix()
m <- mean(matga)

# ESTIMANDO O EFEITO DA INTERAÇÃO
mat_ef <- matrix(NA, ncol = ncol(matga), nrow = nrow(matga))

for(i in 1:nrow(matga)){
  egi <- mean(matga[i, ]) - m
  for(j in 1:ncol(matga)){
    eaj <- mean(matga[, j]) - m
    mat_ef[i, j] <- matga[i, j] - m - egi - eaj
  }
}
colnames(mat_ef) <- colnames(matga)
rownames(mat_ef) <- rownames(matga)


# criar um mapa de calor
dfggp <- 
mat_ef |> 
  as.data.frame() |> 
  tibble::rownames_to_column("gen") |> 
  pivot_longer(-gen, names_to = "amb", values_to = "rg")

# phm <- 
ggplot(dfggp, aes(gen, amb, fill = rg))+
  geom_tile(color = "black") +
  scale_fill_gradient2(low = "red", high = "blue") +
  scale_x_discrete(expand = expansion(0)) +
  scale_y_discrete(expand = expansion(0)) +
  theme(axis.text.x = element_text(angle = 45,
                                   vjust = 1,
                                   hjust = 1)) +
  labs(x = "Genótipo",
       y = "Ano de cultivo")

ggsave("figs/heatmap.jpg",
       width = 8,
       height = 2)


```

# Adaptabilidade
```{r}
mod <- waas_means(df_ggplot, AMB, GEN, RGPLA)
p1 <- 
  plot_scores(mod, type = 2) + 
  theme_stata() +
  coord_equal()
p2 <- 
  plot_scores(mod, type = 3) +
    theme_stata() + 
    labs(x = "Rendimento de grãos por planta (g)",
       y = "Média ponderada dos escores absolutos (WAAS)")

p1 + p2 + 
  plot_annotation(tag_levels = "a") +
  plot_layout(guides = "collect") & theme(legend.position = "bottom")

ggsave("figs/waas_biplot.jpg",
       width = 13,
       height = 7)
```

# Estabilidade
## Método Wricke

WRICKE, G. Zur Berechnung der Ökovalenz bei Sommerweizen und Hafer. Zeitschrift für Pflanzenzüchtung, v.52, p.127-138, 1965.

```{r}
ecovalencia <- 
  apply(mat_ef, 1, function(x){
    sum(x^2)
  })
eco <- data.frame(ecovalencia) |> rownames_to_column("GEN")


```


## Método WAAS
OLIVOTO, T. et al. Mean Performance and Stability in Multi‐Environment Trials I: Combining Features of AMMI and BLUP Techniques. Agronomy Journal, v. 111, n. 6, p. 2949–2960, 2019. 


```{r}
waas_vals <- 
  get_model_data(mod) |> 
  rename(WAAS = RGPLA)

stabs <- left_join(waas_vals, eco)
ggplot(stabs, aes(WAAS, ecovalencia)) +
  geom_point()

cor(stabs$WAAS, stabs$ecovalencia)

```



# Section info
```{r}
sessionInfo()
```
