==> 06_trilha.qmd <==
---
title: "06: Análise de trilha em linhaça"
---

```{r global_options, include = FALSE}
knitr::opts_chunk$set(cache = FALSE,
                      comment = "##",
                      collapse = TRUE,
                      warning = FALSE,
                      message = FALSE)

```


# Pacotes

```{r warning=FALSE, message=FALSE}
library(rio)
library(tidyverse)
library(metan)

planta <- 
  import("data/dados_labfoto.xlsx") |> 
  select(UNIQUE_ID, GEN, PLANTA:NG, -NPT) |> 
  remove_rows_na() |> 
  mutate(MMG = MG / NG * 1000,
         ICC = MG / MC)


ggplot(planta, aes(NG, MG)) +
  geom_point()
# bindar as informações dos grãos

graos <- 
  import("data/graos_raw.csv") |> 
  group_by(img) |> 
  summarise(
    ELGRA = mean(elongation)
  ) |> 
  rename(UNIQUE_ID = img)

capsulas <- 
  import("data/capsulas_raw.csv") |> 
  group_by(img) |> 
  summarise(
    ELCAP = mean(elongation)
  ) |> 
  rename(UNIQUE_ID = img)

dfplantas <- reduce(list(planta, graos, capsulas), left_join)

dfmedias <- 
  dfplantas |> 
  group_by(GEN) |> 
  summarise(across(where(is.numeric), \(x){mean(x, na.rm = TRUE)})) |> 
  remove_rows_na()

```


# Diagnóstico da multicolinearidade
## VIFS
```{r}
colin_plantas <- colindiag(dfplantas, -MG)
colin_medias <- colindiag(dfmedias, - MG)

coldiag <- 
  colin_plantas$VIF |> rename(VIF_PLANTA = VIF) |> 
  left_join(
    colin_medias$VIF |> rename(VIF_MEDIA = VIF)
  )

coldiag


det(as.matrix(colin_medias$cormat))

```

## Autovalores-autovetores
```{r}
evmed <- colin_medias$evalevet |> mutate(cenario = "medias", .before = 1)
evplanta <- colin_plantas$evalevet |> mutate(cenario = "plantas", .before = 1)
evalevet <- bind_rows(evmed, evplanta) |> round_cols(digits = 3)
export(evalevet, "data/evalevet.xlsx")

```

# Trilha
## Sem nenhuma correção


#Regressão em crista
#Diagonal Direto, linha indireto
#DIreto -> efeito puro da variável, soma da linha -> pearson.
#VIF, menor que 10
#Condição(Razão maior e menor abaixo de 100, )
#Determinante quanto menor, mais problemático


## Estimando os coeficientes
```{r}
library(metan)
cor.x <- as.matrix(colin_medias$cormat)
correl <- corr_coef(dfmedias)
cor.y <- as.vector(corr_focus(correl, MG)$MG)

direto <- solve(cor.x %*% t(cor.x)) %*% cor.x %*% cor.y
n <- ncol(cor.x)
Coeff <- data.frame(cor.x)
for (i in 1:n) {
  for (j in 1:n) {
    Coeff[i, j] <- direto[j] * cor.x[i, j]
  }
}
Coeff
```


## Seleção de variáveis
```{r}
trilha2_planta <- 
  path_coeff(dfplantas, 
             resp = MG,
             pred = everything(),
             brutstep = TRUE)

plot(trilha2_planta$Models$Model_1)

export(trilha2_planta$Summary, "data/brutestep.xlsx")

# ggsave("figs/trilha2_planta.jpg",
#        width = 6, height = 6)
```


# Section info
```{r}
sessionInfo()

```
